Solve 2022/Day14 and fix test

+++ src/main/kotlin/aoc2022/Day1.kt
@@ -5,14 +5,14 @@ import splitBy
 import util.Day
 
 class Day1: Day("1") {
-    override fun executePart1(name: String): Int {
+    override fun executePart1(name: String): Long {
        return getInputAsLines(name)
            .splitBy { it.isBlank() }
            .filter { it.isNotEmpty() && it.all { s -> s.isNotEmpty() }}
-           .maxOf { it.sumOf { s -> s.toInt() } }
+           .maxOf { it.sumOf { s -> s.toInt() } }.toLong()
     }
 
-    override fun executePart2(name: String): Int {
+    override fun executePart2(name: String): Long {
         return getInputAsLines(name)
             .splitBy { it.isBlank() }
             .asSequence()
@@ -20,7 +20,7 @@ class Day1: Day("1") {
             .map { it.sumOf { s -> s.toInt() } }
             .sortedDescending()
             .take(3)
-            .sum()
+            .sum().toLong()
     }
 
 }
